+++ stroom-query-common/src/test/java/stroom/query/common/v2/TestConditionalWait.java
@@ -1,6 +1,7 @@
 package stroom.query.common.v2;
 
 import org.assertj.core.api.Assertions;
+import org.junit.Ignore;
 import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -11,6 +12,7 @@ import java.time.temporal.ChronoUnit;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.atomic.AtomicBoolean;
 
+@Ignore
 public class TestConditionalWait {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(TestConditionalWait.class);